Prune old examples and add kitchen sink
