+++ README.md
@@ -2,8 +2,8 @@
 Entity is a Python framework for building AI agents. It runs locally by default and sets up resources automatically.
 
 ## Examples
-Run `python examples/zero_config_agent.py` for a minimal CLI demo or
-`python examples/advanced_workflow.py` to see a multi-stage workflow in action.
+Run `python examples/default_agent.py` for a minimal CLI demo or
+`python examples/kitchen_sink.py` for a full-featured vLLM example.
 Resources are prepared automatically using ``load_defaults()`` so Docker is no
 longer required for these examples.
 The old `[examples]` extra has been removed.
