--- examples/advanced_workflow.yaml
@@ -1,15 +1,0 @@
-input:
-  - entity.plugins.examples.input_reader.InputReader
-parse:
-  - entity.plugins.examples.keyword_extractor.KeywordExtractor
-  - entity.plugins.defaults.ParsePlugin
-think:
-  - entity.plugins.examples.reason_generator.ReasonGenerator
-  - entity.plugins.defaults.ThinkPlugin
-do:
-  - entity.plugins.examples.calculator.Calculator
-review:
-  - entity.plugins.examples.static_reviewer.StaticReviewer
-  - entity.plugins.defaults.ReviewPlugin
-output:
-  - entity.plugins.examples.output_formatter.OutputFormatter
