--- examples/basic_config.yaml
@@ -1,8 +1,0 @@
-resources: {}
-workflow:
-  parse:
-    - entity.plugins.defaults.ParsePlugin
-  think:
-    - entity.plugins.defaults.ThinkPlugin
-  output:
-    - entity.plugins.defaults.OutputPlugin
