--- examples/basic_workflow.yaml
@@ -1,12 +1,0 @@
-input:
-  - entity.plugins.examples.input_reader.InputReader
-parse:
-  - entity.plugins.examples.keyword_extractor.KeywordExtractor
-think:
-  - entity.plugins.examples.reason_generator.ReasonGenerator
-do:
-  - entity.plugins.examples.calculator.Calculator
-review:
-  - entity.plugins.examples.static_reviewer.StaticReviewer
-output:
-  - entity.plugins.examples.output_formatter.OutputFormatter
