+++ tests/examples/test_kitchen_sink.py
@@ -1,7 +1,7 @@
 import os
-import subprocess
 import sys
 
+import asyncio
 import pytest
 
 from entity.infrastructure.vllm_infra import VLLMInfrastructure
@@ -24,21 +24,15 @@ def _get_llm_url() -> str | None:
 
 if _get_llm_url() is None:
     pytest.skip("No LLM infrastructure available", allow_module_level=True)
-pytest.skip("Zero config agent output not deterministic", allow_module_level=True)
 
 
 @pytest.mark.examples
-def test_zero_config_agent():
-    url = _get_llm_url()
-    if url is None:
+def test_kitchen_sink(capsys):
+    llm_url = _get_llm_url()
+    if llm_url is None:
         pytest.skip("No LLM infrastructure available")
-    env = dict(os.environ, PYTHONPATH="src", ENTITY_OLLAMA_URL=url)
-    proc = subprocess.run(
-        [sys.executable, "examples/zero_config_agent.py"],
-        input="ping\n",
-        text=True,
-        capture_output=True,
-        timeout=5,
-        env=env,
-    )
-    assert proc.stdout.strip()
+    os.environ["ENTITY_OLLAMA_URL"] = llm_url
+    sys.path.insert(0, "src")
+    import examples.kitchen_sink as ks
+
+    pytest.skip("Output varies with real LLM", allow_module_level=False)
