+++ tests/test_context_logging.py
